dummy_data/requirements.txt
numpy
pandas
mysql-connector-python
python-dotenv